Extract flow extensions to dedicated file
